Finished validators. oprValidator now in production. Tests started.

+++ 11_CDIO_FINAL/src/final_cdio_11/java/data/validator/IRaavareBatchValidator.java
@@ -1,12 +1,11 @@
 package final_cdio_11.java.data.validator;
 
 import final_cdio_11.java.data.dto.RaavareBatchDTO;
-import final_cdio_11.java.data.dto.RaavareDTO;
 
 public interface IRaavareBatchValidator {
 	boolean isRbIdValid(int rbId);
 	boolean isRaavareIdValid(int raavareId);
-	boolean isAmountValid(int amount);
+	boolean isAmountValid(double amount);
 	boolean isRbStatusValid(int status);
 	boolean isRbValid(RaavareBatchDTO rbDTO);
 }
